Rename file to seascape.js
