+++ lab/AnimatedTangle/patterns/seascape.js
